// models/saftAgreementData.ts
interface SaftAgreementData {
  customTitle: string;
  partyName: string;
  partyEmail: string;
  partyAddress: string;
  partyWallet: string;
  date: string;
  counterPartyName: string;
  counterPartyEmail: string;
  counterPartyAddress: string;
  counterPartyWallet: string;
  purchaseAmount: number;
  jurisdiction: string;
  typeOfCompany: string;
  tokenAmount: number;
  discountRate: number;
  website: string;
  paymentOption: string;
  bankName: string;
  address: string;
  aba: string;
  payeeAccount: string;
  payeeAccountName: string;
  ethereum: string;
  bitcoin: string;
}

export default SaftAgreementData;